+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+- feat(init): add interactive provider detection with Azure DevOps prompts and `--non-interactive` GitHub defaults (#468)
+
 ## [0.23.0] - 2026-05-04
 
 ### Added